Add a new category for unit-testing tools and libraries
This patch creates a new "Testing" category and populates it with
links to a variety of unit-testing tools and libraries.  These links
could arguably go into the "Utilities" category, but there are a few
problems with this:

1. Unit-testing tools are a specific type of utility and thus,
   arguably, deserve their own category.

2. Having all unit-testing libraries under "Utilities" may lead to
   future bloat.

Some existing links in the list may warrant being moved into "Testing"
but this patch does not address that and does not move any existing
link from its current category.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #34