Add the existing 'knife' library to the "Testing" category
We already have 'knife' listed elsewhere.  However, the library
provides the 'knife.test' sub-module specifically for unit-testing.
This patch adds a link directly to that sub-module into the "Testing"
category, without moving the existing link to 'knife' since the
library serves a lot of purposes, and moving it exclusively into
"Testing" would not be appropriate.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #34