Added brinevector to Math category
While looking for vector libraries for lua, I noticed most of them use tables to store the vectors themselves. This might be fine for most applications, but for games and high-performance uses, creating a table for every vector is simply too much overhead. Using C data to store and create vectors with the `ffi` library in luajit is a much more efficient method that can produce code that performs much faster and consumes a lot less memory.

So I wrote this vector library to take advantage of that fact, and made it extremely beginner-friendly and easy for everybody to use. It's also designed to be lightweight and very portable; it's only really a single file!

I know CPML also uses ffi, but mine is different enough in that it's lightweight and standalone. And getting some properties from vectors in mine is done as a simple . access instead of a method (ie, v.length vs v:length() )