Add parallax
parallax is a camera utility library for LÖVE that adds parallax
scrolling when used with a camera library.

You can create layers and a layer can draw an image with perfect tiling
on x, y, or both axes.

parallax is not itself a camera system! I use parallax in my projects to
add parallax scrolling to gamera.

[Here's a demo](https://idbrii.github.io/love-parallax/demo/) showing
drawing on multiple layers and a tiled background image.